impl changes and structure changes
Focusing on the speed. Chaged the implementation of _eval_teams and
merging some sub evals that always go together to one private method.

The LID finding is rather unnicely implemented and is slow and
unnecessarily complex (humpolácky). This should be the part to change if the data
management is too slow.

To be continued (soon hopefully)...